Gui: split selection stack implementation
